requirements.txt
# Main dependencies
SQLAlchemy
sqlalchemy-utils
PyMySQL
pyyaml

# Test dependencies
pytest
tox
coverage
python-coveralls
